WeatherForm moved to settings; Settings slice
