+++ src/features/emotionsCheckin/EmotionsCheckin.module.css
@@ -49,7 +49,7 @@ img{
 
 
 
-button{
+.button{
   appearance: none;
   background: none;
   color: transparent;
